fix: configs
